requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
passlib
bcrypt==3.2.0
pyjwt
python-dotenv
redis
slowapi
email-validator
multipart
psycopg2-binary
alembic
